+++ src/main/kotlin/com/github/crunchwrap89/featureorchestrator/startup/MyProjectActivity.kt
@@ -1,8 +1,12 @@
 package com.github.crunchwrap89.featureorchestrator.startup
 
+import com.github.crunchwrap89.featureorchestrator.featureorchestrator.skills.SkillService
+import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.ProjectActivity
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 
 class MyProjectActivity : ProjectActivity {